+++ versions.tf
@@ -8,7 +8,3 @@ terraform {
   }
   required_version = ">= 1.5.0"
 }
-
-provider "aws" {
-  region = "eu-west-2"
-}