+++ CHANGELOG.md
@@ -61,3 +61,13 @@ eg:
 ### [0.0.15] 2026-01-20
 
   * Updating tags, adding RDS variables, secret rotation functionality and security group logic
+
+### [0.0.16] 2026-01-20
+
+  * Adding RDS variables and secret rotation functionality
+
+
+### [0.0.17] 2026-01-21
+
+  * Adding RDS variable - max_allocated_storage
+